#330 moved event filters to the directory EventSearch
