+++ frontend/src/Views/SearchEvents/EventSearch/FilterComponents/helper.js
